requirements.txt
asgiref==3.5.1
attrs==22.1.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
Django==4.0.4
django-cors-headers==3.13.0
django-mongodb-engine==0.6.0
django-mongodb-engine-py3==0.6.0.1
django-phonenumber-field==6.1.0
django-rest-swagger==2.1.2
djangorestframework==3.13.1
djangotoolbox==1.8.0
djongo==1.3.6
dnspython==2.2.1
drf-yasg==1.21.3
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.0
jsonschema==4.14.0
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
phonenumbers==8.12.48
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
pymongo==3.12.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-http-client==3.3.7
pytz==2022.1
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.2.4
starkbank-ecdsa==2.0.3
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
waitress==2.1.1
whitenoise==6.2.0
